requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
albumentations==1.2.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.8
cmake==3.25.0
contextlib2==21.6.0
contourpy==1.3.0
cycler==0.12.1
einops==0.8.1
eval_type_backport==0.2.2
filelock==3.13.1
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.73.1
h5py==3.14.0
hf-xet==1.1.5
huggingface-hub==0.17.3
idna==3.4
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.4
joblib==1.5.1
kiwisolver==1.4.7
lightning==2.5.2
lightning-utilities==0.14.3
lit==15.0.7
Markdown==3.8.2
MarkupSafe==2.1.5
matplotlib==3.9.4
MedPy==0.5.2
ml_collections==0.1.1
monai==1.2.0
mpmath==1.3.0
multidict==6.6.3
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.6.77
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
packaging==25.0
pandas==2.3.1
pillow==11.0.0
platformdirs==4.3.8
propcache==0.3.2
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.2
pytorch-wavelets==1.3.0
pytz==2025.2
PyWavelets==1.6.0
PyYAML==6.0.2
qudida==0.0.4
regex==2024.11.6
requests==2.28.1
safetensors==0.5.3
scikit-image==0.18.3
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
sentry-sdk==2.33.0
simpleitk==2.5.2
six==1.17.0
smmap==5.0.2
sympy==1.13.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2024.8.30
timm==0.9.2
tokenizers==0.14.1
torch==2.7.0+cu118
torchaudio==2.7.0+cu118
torchmetrics==1.7.4
torchsummary==1.5.1
torchvision==0.22.0+cu118
tqdm==4.67.1
transformers==4.34.1
triton==3.3.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==1.26.13
wandb==0.21.0
Werkzeug==3.1.3
wget==3.2
yacs==0.1.8
yarl==1.20.1
zipp==3.23.0
